tidying up the code
